pyproject.toml
[project]
name = "maxvoice"
version = "0.1.0"
description = "Typeless-like voice dictation with pluggable Azure AI Foundry models"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "PyQt6>=6.7",
    "sounddevice>=0.4.7",
    "numpy>=2.0",
    "scipy>=1.13",
    "pynput>=1.7.7",
    "openai>=1.55",
    "pyperclip>=1.9",
    "pyautogui>=0.9.54",
    "sqlmodel>=0.0.22",
    "python-dotenv>=1.0",
    "pydantic>=2.9",
    "httpx>=0.27",
]

[project.scripts]
maxvoice = "maxvoice.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/maxvoice"]

[tool.uv]
dev-dependencies = []
